Add join model
This commit will add a join model for reviews. It will create a one to many association of move to review and one to many association for user to review.
It will also update all necessary tempalates to display all reviews per user.
It will modify the comments form to remove name field, since only logged in user can
add a comment, hence the name will be automatically populated.