Add Genre feature
This commit will add the Genre feature to the app. It will add
the characterization through model and create the many-to-many
association between movie and genre via that model.

It adds checkbox feature for admins to add genres to a movie.
It will be part of the move_params hash.
Edit movie is also updated with this feature