Add custom scopes
This commit will add custom scopes / queries to the app. It will use
ruby scope meethod with lambdas to achieve this.

The link_to method in the header is conveted in a custom nav_link_to
method in order to underline / use the css active class for the active link

New links are added in the header nav.