Ready for deployment
The issues discovered in previous commit, see commit note, are resolved.
Restart of the server has fixed the issues, so no code changes.

This commit will add the necessary changes to deploy the app on heroku.
It will add the pg adapter for heroku and add linux as plattform to the
gemfile.lock